hackathon/agent/requirements.txt
# AI Multi-Tab Browser Agent — Dependencies
# Install: pip install -r requirements.txt

# Browser automation
playwright>=1.42.0

# AI / Gemini
google-genai>=1.0.0

# WebSocket streaming
websockets>=12.0

# HTML parsing
beautifulsoup4>=4.12.0

# Playwright browser binaries (run separately):
# playwright install chromium
